requirements.txt
Adafruit-Blinka==2.3.1
adafruit-circuitpython-neopixel==3.3.7
Adafruit-PlatformDetect==1.3.2
Adafruit-PureIO==0.2.3
requests==2.22.0
rpi-ws281x==4.2.2
RPi.GPIO==0.7.0
